Refactor fragments
