+++ app/src/main/java/com/sdpteam/connectout/fragments/HomeFragment.java
@@ -1,5 +1,6 @@
-package com.sdpteam.connectout.ui.home;
+package com.sdpteam.connectout.fragments;
 
+import android.annotation.SuppressLint;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -8,13 +9,20 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
+import com.sdpteam.connectout.R;
 import com.sdpteam.connectout.databinding.FragmentHomeBinding;
+
+import java.util.zip.Inflater;
 
 public class HomeFragment extends Fragment {
 
     private FragmentHomeBinding binding;
+
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
@@ -26,12 +34,26 @@ public class HomeFragment extends Fragment {
 
         final TextView textView = binding.textHome;
         homeViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
-        return root;
+        return binding.getRoot();
     }
 
     @Override
     public void onDestroyView() {
         super.onDestroyView();
         binding = null;
+    }
+
+    public static class HomeViewModel extends ViewModel {
+
+        private final MutableLiveData<String> mText;
+
+        public HomeViewModel() {
+            mText = new MutableLiveData<>();
+            mText.setValue("This is home fragment");
+        }
+
+        public LiveData<String> getText() {
+            return mText;
+        }
     }
 }
