+++ app/src/main/java/com/sdpteam/connectout/fragments/MyAccountFragment.java
@@ -1,11 +1,13 @@
-package com.sdpteam.connectout.ui.myAccount;
+package com.sdpteam.connectout.fragments;
 
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
 import androidx.lifecycle.ViewModelProvider;
 
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
@@ -13,7 +15,6 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import com.sdpteam.connectout.R;
 import com.sdpteam.connectout.databinding.FragmentMyAccountBinding;
 
 
@@ -40,4 +41,16 @@ public class MyAccountFragment extends Fragment {
         binding = null;
     }
 
+    public static class MyAccountViewModel extends ViewModel {
+        private final MutableLiveData<String> mText;
+
+        public MyAccountViewModel() {
+            mText = new MutableLiveData<>();
+            mText.setValue("This is my account fragment");
+        }
+
+        public LiveData<String> getText() {
+            return mText;
+        }
+    }
 }
