--- app/src/main/java/com/sdpteam/connectout/ui/filter/FilterViewModel.java
@@ -1,19 +1,0 @@
-package com.sdpteam.connectout.ui.filter;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class FilterViewModel extends ViewModel {
-
-    private final MutableLiveData<String> mText;
-
-    public FilterViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is filter fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
