--- app/src/main/java/com/sdpteam/connectout/ui/myAccount/MyAccountViewModel.java
@@ -1,18 +1,0 @@
-package com.sdpteam.connectout.ui.myAccount;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class MyAccountViewModel extends ViewModel {
-    private final MutableLiveData<String> mText;
-
-    public MyAccountViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is my account fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
