--- app/src/main/java/com/sdpteam/connectout/ui/myEvents/MyEventsViewModel.java
@@ -1,19 +1,0 @@
-package com.sdpteam.connectout.ui.myEvents;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class MyEventsViewModel extends ViewModel {
-
-    private final MutableLiveData<String> mText;
-
-    public MyEventsViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is my event fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
