+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.HomeFragment"
+    tools:context=".fragment.HomeFragment"
+    android:id="@+id/home_fragment"
     android:background="@color/purple_500">
 
     <TextView
