+++ app/src/main/res/layout/fragment_my_account.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.myAccount.MyAccountFragment"
+    tools:context=".fragments.MyAccountFragment"
     android:background="@color/teal_200">
 
     <TextView
